Refactor: User의 당첨결과 값 객체화
User내에 존재하는 profit, resultMatch 변수를 하나의 WinningStatic으로 객체화
